requirements.txt
# Construction Estimation App Requirements
# Consolidated from original requirements.txt and requirements_estimation.txt

# Core dependencies
aiofiles==24.1.0
annotated-types==0.7.0
anthropic==0.57.1
anyio==4.9.0
Brotli==1.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
defusedxml==0.7.1
distro==1.9.0

# Note: OCR dependencies removed

# Web framework and UI
fastapi==0.116.1
ffmpy==0.6.0
filelock==3.18.0
fonttools==4.59.0
fpdf2==2.8.3
fsspec==2025.7.0
gradio==5.38.0
gradio_client==1.11.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.4

# Database (for construction estimation)
sqlalchemy==2.0.23
alembic==1.12.1

# Data processing
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jiter==0.10.0
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
openai==1.97.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.11.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0

# Development tools
ruff==0.12.3
pytest==7.4.3
black==23.12.1
flake8==6.1.0
mypy==1.8.0

# Other utilities
safehttpx==0.1.6
scikit-image==0.25.2
scipy==1.16.0
semantic-version==2.10.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.1
sympy==1.14.0
tifffile==2025.6.11
tomlkit==0.13.3
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1

# Optional: For PostgreSQL support later
# psycopg2-binary==2.9.9

# Note: EasyOCR support removed
